+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,10 +2,13 @@
 
 Release History
 ===============
+0.1.30
+++++++
+* webapp: az webapp update supports httpsOnly
+
 0.1.29
 ++++++
 * webapp/functionapp: author managed identity commands `identity assign/show`, and deprecate `assign-identity`
-* webapp: az webapp update supports httpsOnly
 
 0.1.28
 ++++++